apps/pipeline/pyproject.toml
[project]
name = "autowork-pipeline"
version = "0.1.0"
description = "AutoWork AI 콘텐츠 자동화 파이프라인"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.36.0",
    "httpx>=0.27.0",
    "supabase>=2.7.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.0",
    "google-api-python-client>=2.145.0",
    "google-auth-oauthlib>=1.2.1",
    "pytrends>=4.9.2",
    "tenacity>=9.0.0",
    "structlog>=24.4.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "apscheduler>=3.10.0",
    # CMP-74: TTS + 영상 생성 파이프라인
    "google-cloud-texttospeech>=2.17.0",
    "Pillow>=10.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
